_internship-tasks/task_6_v2/requirements.txt
pyspark==3.5.2
delta-spark==3.2.0